requirements.txt
pyzmq==25.1.2
sounddevice==0.4.6
numpy